vic: emulate bad lines in top border
Full emulation is currently skipped for border lines, but it must be
done for (the first) bad line which can be in the border due to YSCROLL.
Otherwise the char/bitmap data is not setup properly and corrupts the
visible lines, as was happening in River Raid. The cycle counting would
be wrong too.

So make an exception for bad lines in the border and process the mode
fully. Afterwards discard the produced pixels and ignore sprite-data
collisions, as is specified for the border.