vic: setting a bad line will not cancel idle in the right border
The program might write a YSCROLL value that makes the current line a
bad line, which also sets idle to 0. In River Raid the bottom HUD was
glitching because of such an YSCROLL write. When the write happened
after idle was set to 1 due to RC being 7 (evaluated before the right
border), it would put idle back to 0 and the next line would be active
and bump the VC, making the whole HUD render a row too early.

This can be prevented by ignoring new bad line condition late in the
line for setting idle back to 0. I'm not sure if it's correct, the doc
can be interpreted either way. Perhaps with more accurate emulation of
cycles, the YSCROLL setting would happen either before the border, thus
not undoing the idle setting to 1, or in the next line, where it would
not cause it to be a bad line in the first place. But for now, this
seems to do the job.