vic: better handling of !CSEL and XSCROLL
With !CSEL, extend left border with remembered color, and also extend
right border. Flip the wrong condition as to when sprites could be
visible or not.
With XSCROLL > 0 bump also 'pe' to let the lines be rendered fully which
should be more precise than inserting clock(1) and removing an
iteration.
Make linebuffer and spriteLine larger to prevent overflow due to this.
(Actually make linebuffer smaller as the *16 was wrong)

We handle neither CSEL flipping for hyperscreen, nor XSCROLL changes
within a line. But we handle CSEL changing in the middle so left/right
sides can differ.