vic: more accurate idle state transition
This should be more according to the doc.

- should not go idle when badline is true, check it
- don't evaluate is_badline() for bumping rc, that's now redundant

This wasn't enough to solve River Raid HUD glitch, so keep the
badLineLate check. But it's very likely a workaround for imprecise
timing, the doc doesn't seem to support it.

FLI section suggests VCBASE also shouldn't reset with badline, but
attempting that also caused River Raid HUD glitch (in the opposite
direction?).

Current theory is due to imprecise timing, River Raid will change
YSCROLL (causing a badline) sometimes in the right border (mitigated by
badlineLate) but sometimes also before the right border (migitated by
resetting VCBASE anyway). If timing was precise enough that YSCROLL
change happened on the beginning of the next line, not triggering
badline, there would be no need for these workarounds.